Write a number-guessing game in which the computer selects a random number in the range 1 to 100, and users get a maximum
2. Write a number-guessing game in which the computer selects a random number in the range 1 to 100, and users get a maximum
of 20 attempts to guess it.
After each guess there should be some hint given to the player as to how close that guess was. If the guess was only 1 away print
a hint like "That guess was hot". If the guess was within 5 points (high or low) print a hint like "You're getting warm". If the guess
was within 10 points print a hint like "Your guess is still cool". If the guess was more than 10 points away print a hint like "You're
guess is cold".
At the end of each game, users should be told whether they won or lost, and how many guesses it took if they got it correct. To
make the game more interesting, the program should vary the wording of the messages that it outputs for winning, for losing, and
for asking for another game. Create as many as 5 different messages for each of these cases and use random numbers to choose
among them. You should put this functionality in two functions called SayYouWon and SayYouLost..........